Fixed issues in Delete History and Delete histories Actions from EquipmentHistoryDashboard

+++ Application/Controllers/EquipmentHistoryDashboard/Actions/DeleteEquipmentHistories_Post.php
@@ -27,7 +27,6 @@ class DeleteEquipmentHistories_Post extends Action {
         $this->qbtxlineidCollection = $this->Request->hasProperty('qbtxlineidCollection') ? json_decode($this->Request->qbtxlineidCollection) : array();
         $this->equipidCollection = $this->Request->hasProperty('equipidCollection') ? json_decode($this->Request->equipidCollection) : array();
 
-
         $this->status = EQUIPMENT_HISTORY_DASHBORD_STATUS_AVAILABLE;
         $this->ordnum = '';
         $this->vesselid = '';
@@ -37,16 +36,18 @@ class DeleteEquipmentHistories_Post extends Action {
 
         foreach ($this->qbtxlineidCollection as $index => $qbtxlineid){
             $currentEquipid = $this->equipidCollection[$index];
-            $entity = $this->controller->DatUnitOfWork->SWEQUIPDRepository->GetByQbtxlineid($qbtxlineid);
-
-            $isSuccess = $this->controller->DatUnitOfWork->SWEQUIPDRepository->Delete($entity);
-            $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->UpdateStatus($currentEquipid, $this->status);
-            $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->UpdateDates($currentEquipid, $this->installdte, $this->expdtein, $this->daterec);
-            $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->RemoveLastHistoryReference($currentEquipid);
-            $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->RemoveLastWorkOrder($currentEquipid);
-            $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->RemoveLastVesselid($currentEquipid);
+            if ($qbtxlineid !== ''){
+                $entity = $this->controller->DatUnitOfWork->SWEQUIPDRepository->GetByQbtxlineid($qbtxlineid);
+                if ($entity) {
+                    $isSuccess = $this->controller->DatUnitOfWork->SWEQUIPDRepository->Delete($entity);
+                    $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->UpdateStatus($currentEquipid, $this->status);
+                    $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->UpdateDates($currentEquipid, $this->installdte, $this->expdtein, $this->daterec);
+                    $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->RemoveLastHistoryReference($currentEquipid);
+                    $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->RemoveLastWorkOrder($currentEquipid);
+                    $isSuccess &= $this->controller->DatUnitOfWork->SWEQUIPRepository->RemoveLastVesselid($currentEquipid);
+                }
+            }
         }
-
 
         if ($isSuccess) {
             $result['success'] = true;
